requirements.txt
torch
torchvision
py-telegram-bot
matplotlib
Pillow
torchinfo
tqdm